requirements.txt
aiohttp==3.7.4.post0
discord==1.0.1
discord.py==1.7.1
numexpr==2.7.1
python-dotenv==0.19.2
virtualenv==20.10.0
requests==2.23.0